docker-compose: adicionando frontend e dockerfiles

+++ Frontend/TesteFrontENd/src/app/deleting-user/deleting-user.ts
@@ -1,17 +1,31 @@
-import { Component, Inject, Input, input } from '@angular/core';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { UserService } from '../service/user-service';
 
 @Component({
   selector: 'app-deleting-user',
   imports: [],
   templateUrl: './deleting-user.html',
-  styleUrl: './deleting-user.css',
+  styleUrls: ['./deleting-user.css'],
 })
 export class DeletingUser {
   @Input() userId: string = '';
-  userService: UserService = Inject(UserService);
-  
-  deleteUser() {
-    this.userService.deleteUser(this.userId);
+
+  constructor(private userService: UserService) {}
+
+  @Output() deleteEvent = new EventEmitter<void>();
+
+  async deleteUser() {
+    if (!this.userId) {
+      return;
+    }
+    try {
+      await this.userService.deleteUser(this.userId);
+      alert("User with ID " + this.userId + " deleted successfully.");
+      this.deleteEvent.emit();
+    } catch {
+      alert("An error occurred while deleting the user!");
+    }
   }
+
+
 }
